refactor: move repo funcs inside db
