+++ package.json
@@ -51,13 +51,13 @@
         "node": ">=22.3.0"
     },
     "dependencies": {
-        "better-sqlite3": "^12.1.1",
+        "better-sqlite3": "^12.2.0",
         "compression": "^1.8.0",
         "connect-flash": "^0.1.1",
         "connect-session-knex": "^5.0.0",
         "cors": "^2.8.5",
         "csrf-sync": "^4.2.1",
-        "dotenv": "^16.6.0",
+        "dotenv": "^17.0.1",
         "ejs": "^3.1.10",
         "express": "^5.1.0",
         "express-ejs-layouts": "^2.5.1",
@@ -68,10 +68,10 @@
         "jsonwebtoken": "^9.0.2",
         "knex": "^3.1.0",
         "marked": "^16.0.0",
-        "nodemailer": "^7.0.3"
+        "nodemailer": "^7.0.4"
     },
     "devDependencies": {
-        "@playwright/test": "^1.53.1",
+        "@playwright/test": "^1.53.2",
         "@types/compression": "^1.8.1",
         "@types/connect-flash": "^0.0.40",
         "@types/cors": "^2.8.19",
@@ -80,10 +80,10 @@
         "@types/express-ejs-layouts": "^2.5.4",
         "@types/express-session": "^1.18.2",
         "@types/jsonwebtoken": "^9.0.10",
-        "@types/node": "^24.0.4",
+        "@types/node": "^24.0.10",
         "@types/nodemailer": "^6.4.17",
-        "@typescript-eslint/eslint-plugin": "^8.35.0",
-        "@typescript-eslint/parser": "^8.35.0",
+        "@typescript-eslint/eslint-plugin": "^8.35.1",
+        "@typescript-eslint/parser": "^8.35.1",
         "@vitest/coverage-v8": "^3.2.4",
         "@wajeht/express-templates-reload": "^2.0.2",
         "eslint-config-prettier": "^10.1.5",
