+++ src/handler.test.ts
@@ -1,6 +1,6 @@
 import { db } from './db/db';
+import { notes } from './db/db';
 import { logger } from './utils/logger';
-import { notes } from './repository';
 import type { Request, Response } from 'express';
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
 import { getHealthzHandler, postExportDataHandler, toggleNotePinHandler } from './handler';
