+++ src/middleware.test.ts
@@ -1,7 +1,7 @@
 import { db } from './db/db';
+import { users } from './db/db';
 import type { User } from './type';
 import { logger } from './utils/logger';
-import { users } from './repository';
 import { Session } from 'express-session';
 import { api, getApiKey, isApiRequest } from './utils/util';
 import type { Request, Response, NextFunction } from 'express';
@@ -51,8 +51,8 @@ vi.mock('./config', () => ({
     },
 }));
 
-vi.mock('./repository', async () => {
-    const actual = await vi.importActual('./repository');
+vi.mock('./db/db', async () => {
+    const actual = await vi.importActual('./db/db');
     return {
         ...(actual as any),
         users: {
